Rename bootstrap package & remove waitForServer
