+++ pkg/spoke/spokebootstrap/bootstrap_test.go
@@ -1,4 +1,4 @@
-package bootstrap
+package spokebootstrap
 
 import (
 	"context"
