+++ pkg/spoke/spokebootstrap/certificate.go
@@ -1,4 +1,4 @@
-package bootstrap
+package spokebootstrap
 
 import (
 	"bytes"
