+++ pkg/spoke/spokebootstrap/clientconfig.go
@@ -1,4 +1,4 @@
-package bootstrap
+package spokebootstrap
 
 import (
 	"context"
