+++ pkg/spoke/spokebootstrap/options.go
@@ -1,4 +1,4 @@
-package bootstrap
+package spokebootstrap
 
 import (
 	"io/ioutil"
